% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_of_graphs.R
\name{types_of_graphs}
\alias{types_of_graphs}
\title{Generate plot of types of graphs matrix}
\usage{
types_of_graphs(code = c("n", "y"))
}
\arguments{
\item{code}{If yes (\code{"y"}), the function will print the code to generate the ggplot2 object in the console. If no (\code{"n"}), the function will return a ggplot2 object.}
}
\value{
If \code{code} is \code{"n"}, a ggplot2 object. If \code{code} is \code{"y"}, nothing (the code is printed in the console).
}
\description{
Generate plot of types of graphs matrix
}
